[Chore] #40 OOTDResponseDTO 하나로 통합
