--- src/main/java/com/nova/yeobaek/domain/ootd/dto/response/CreateOOTDResponse.java
@@ -1,4 +1,0 @@
-package com.nova.yeobaek.domain.ootd.dto.response;
-
-/** OOTD 등록 응답 DTO */
-public record CreateOOTDResponse(Long ootdId) {}
