--- src/main/java/com/nova/yeobaek/domain/ootd/dto/response/OOTDItemDetailResponse.java
@@ -1,16 +1,0 @@
-package com.nova.yeobaek.domain.ootd.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class OOTDItemDetailResponse {
-
-    private Long fashionItemId;
-    private float posX;
-    private float posY;
-    private float scale;
-    private float rotation;
-    private int zIndex;
-}
